backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
SQLAlchemy==2.0.30
pymysql==1.1.0    
python-dotenv==1.0.1
pandas==2.0.3
requests==2.31.0
